+++ app/controllers/premiumLease/PremiumsGrantLeaseController.scala
@@ -44,7 +44,7 @@ class PremiumsGrantLeaseController @Inject()(
 
   val form = formProvider()
 
-  def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData) {
+  def onPageLoad(taxYear: Int, mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData) {
     implicit request =>
 
       val preparedForm = request.userAnswers.get(PremiumsGrantLeasePage) match {
@@ -52,21 +52,21 @@ class PremiumsGrantLeaseController @Inject()(
         case Some(value) => form.fill(value)
       }
 
-      Ok(view(preparedForm, mode))
+      Ok(view(preparedForm, taxYear, mode))
   }
 
-  def onSubmit(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData).async {
+  def onSubmit(taxYear: Int, mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData).async {
     implicit request =>
 
       form.bindFromRequest().fold(
         formWithErrors =>
-          Future.successful(BadRequest(view(formWithErrors, mode))),
+          Future.successful(BadRequest(view(formWithErrors,taxYear, mode))),
 
         value =>
           for {
             updatedAnswers <- Future.fromTry(request.userAnswers.set(PremiumsGrantLeasePage, value))
             _              <- sessionRepository.set(updatedAnswers)
-          } yield Redirect(navigator.nextPage(PremiumsGrantLeasePage, mode, updatedAnswers))
+          } yield Redirect(navigator.nextPage(PremiumsGrantLeasePage, taxYear ,mode, updatedAnswers))
       )
   }
 }
