+++ app/forms/premiumLease/CalculatedFigureYourselfFormProvider.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package forms
-
-import javax.inject.Inject
+package forms.premiumLease
 
 import forms.mappings.Mappings
 import play.api.data.Form
+
+import javax.inject.Inject
 
 class CalculatedFigureYourselfFormProvider @Inject() extends Mappings {
