+++ app/views/premiumLease/CalculatedFigureYourselfView.scala.html
@@ -14,6 +14,8 @@
  * limitations under the License.
  *@
 
+@import controllers.premiumLease.routes._
+
 @this(
     layout: templates.Layout,
     formHelper: FormWithCSRF,
@@ -26,7 +28,7 @@
 
 @layout(pageTitle = title(form, messages("calculatedFigureYourself.title"))) {
 
-    @formHelper(action = routes.CalculatedFigureYourselfController.onSubmit(taxYear, mode), Symbol("autoComplete") -> "off") {
+    @formHelper(action = CalculatedFigureYourselfController.onSubmit(taxYear, mode), Symbol("autoComplete") -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
