+++ app/views/premiumLease/PremiumsGrantLeaseView.scala.html
@@ -26,7 +26,7 @@
     govukButton: GovukButton
 )
 
-@(form: Form[_], mode: Mode)(implicit request: Request[_], messages: Messages)
+@(form: Form[_],taxYear: Int ,mode: Mode)(implicit request: Request[_], messages: Messages)
 
 @layout(pageTitle = title(form, messages("premiumsGrantLease.title"))) {
 
@@ -65,7 +65,7 @@
     </p>
 
 
-@formHelper(action = PremiumsGrantLeaseController.onSubmit(mode), Symbol("autoComplete") -> "off") {
+@formHelper(action = PremiumsGrantLeaseController.onSubmit(taxYear, mode), Symbol("autoComplete") -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
