+++ test/controllers/premiumLease/CalculatedFigureYourselfControllerSpec.scala
@@ -14,48 +14,49 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.premiumLease
 
 import base.SpecBase
-import forms.premiumLease.LeasePremiumPaymentFormProvider
+import controllers.premiumLease.routes
+import forms.premiumLease.CalculatedFigureYourselfFormProvider
 import models.{NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.premiumLease.LeasePremiumPaymentPage
+import pages.CalculatedFigureYourselfPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
-import views.html.LeasePremiumPaymentView
+import views.html.CalculatedFigureYourselfView
 
 import java.time.LocalDate
 import scala.concurrent.Future
 
-class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
+class CalculatedFigureYourselfControllerSpec extends SpecBase with MockitoSugar {
 
-  def onwardRoute : Call = Call("GET", "/foo")
+  def onwardRoute = Call("GET", "/foo")
 
-  val formProvider = new LeasePremiumPaymentFormProvider()
+  val formProvider = new CalculatedFigureYourselfFormProvider()
   val form = formProvider()
-  private val taxYear = LocalDate.now.getYear
+  val taxYear = LocalDate.now.getYear
 
-  lazy val leasePremiumPaymentRoute = routes.LeasePremiumPaymentController.onPageLoad(taxYear, NormalMode).url
+  lazy val calculatedFigureYourselfRoute = routes.CalculatedFigureYourselfController.onPageLoad(taxYear, NormalMode).url
 
-  "LeasePremiumPayment Controller" - {
+  "CalculatedFigureYourself Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
-      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), isAgent = false).build()
+      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), false).build()
 
       running(application) {
-        val request = FakeRequest(GET, leasePremiumPaymentRoute)
+        val request = FakeRequest(GET, calculatedFigureYourselfRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[LeasePremiumPaymentView]
+        val view = application.injector.instanceOf[CalculatedFigureYourselfView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, taxYear, NormalMode)(request, messages(application)).toString
@@ -64,14 +65,14 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val userAnswers = UserAnswers(userAnswersId).set(LeasePremiumPaymentPage, true).success.value
+      val userAnswers = UserAnswers(userAnswersId).set(CalculatedFigureYourselfPage, true).success.value
 
-      val application = applicationBuilder(userAnswers = Some(userAnswers), isAgent = false).build()
+      val application = applicationBuilder(userAnswers = Some(userAnswers), false).build()
 
       running(application) {
-        val request = FakeRequest(GET, leasePremiumPaymentRoute)
+        val request = FakeRequest(GET, calculatedFigureYourselfRoute)
 
-        val view = application.injector.instanceOf[LeasePremiumPaymentView]
+        val view = application.injector.instanceOf[CalculatedFigureYourselfView]
 
         val result = route(application, request).value
 
@@ -87,7 +88,7 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
       when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
 
       val application =
-        applicationBuilder(userAnswers = Some(emptyUserAnswers), isAgent = false)
+        applicationBuilder(userAnswers = Some(emptyUserAnswers), false)
           .overrides(
             bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
             bind[SessionRepository].toInstance(mockSessionRepository)
@@ -96,7 +97,7 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, leasePremiumPaymentRoute)
+          FakeRequest(POST, calculatedFigureYourselfRoute)
             .withFormUrlEncodedBody(("value", "true"))
 
         val result = route(application, request).value
@@ -108,16 +109,16 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must return a Bad Request and errors when invalid data is submitted" in {
 
-      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), isAgent = false).build()
+      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), false).build()
 
       running(application) {
         val request =
-          FakeRequest(POST, leasePremiumPaymentRoute)
+          FakeRequest(POST, calculatedFigureYourselfRoute)
             .withFormUrlEncodedBody(("value", ""))
 
         val boundForm = form.bind(Map("value" -> ""))
 
-        val view = application.injector.instanceOf[LeasePremiumPaymentView]
+        val view = application.injector.instanceOf[CalculatedFigureYourselfView]
 
         val result = route(application, request).value
 
@@ -128,10 +129,10 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must redirect to Journey Recovery for a GET if no existing data is found" in {
 
-      val application = applicationBuilder(userAnswers = None, isAgent = true).build()
+      val application = applicationBuilder(userAnswers = None, true).build()
 
       running(application) {
-        val request = FakeRequest(GET, leasePremiumPaymentRoute)
+        val request = FakeRequest(GET, calculatedFigureYourselfRoute)
 
         val result = route(application, request).value
 
@@ -142,11 +143,11 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must redirect to Journey Recovery for a POST if no existing data is found" in {
 
-      val application = applicationBuilder(userAnswers = None, isAgent = true).build()
+      val application = applicationBuilder(userAnswers = None, true).build()
 
       running(application) {
         val request =
-          FakeRequest(POST, leasePremiumPaymentRoute)
+          FakeRequest(POST, calculatedFigureYourselfRoute)
             .withFormUrlEncodedBody(("value", "true"))
 
         val result = route(application, request).value
