+++ test/forms/propertyrentals/PremiumsGrantLeaseFormProviderSpec.scala
@@ -1,4 +1,4 @@
-package forms
+package forms.propertyrentals
 
 import forms.behaviours.IntFieldBehaviours
 import forms.premiumLease.PremiumsGrantLeaseFormProvider
